refactor(webapp): move InputOTP out of components-v2
components-v2 was retired months ago; this file was placed there by a
recent PR whose bot review comment cited stale AGENTS.md guidance.
